requirements.txt
Django>=4.0
djangorestframework>=3.14
torch>=2.0
torchvision>=0.15
tensorflow>=2.12
pillow>=9.0
numpy
